# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Event-driven notification system assessment API.",
        "title": "Insider One Assessment Notification API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:18080",
    "basePath": "/",
    "paths": {
...
        "/templates": {
            "get": {
                "description": "Returns all templates.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "List templates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template channel",
                        "name": "channel",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_dto_resource.TemplatesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.HTTPErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.HTTPErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a message template that can be referenced by notifications.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Create template",
                "parameters": [
                    {
                        "description": "Create template request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_dto_request.CreateTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_dto_resource.TemplateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.HTTPErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.HTTPErrorResponse"
                        }
                    }
                }
            }
        },
        "/templates/{id}": {
            "get": {
                "description": "Returns a template by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Get template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_dto_resource.TemplateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.HTTPErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.HTTPErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.HTTPErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates a template by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Update template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update template request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_dto_request.UpdateTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_dto_resource.TemplateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.HTTPErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.HTTPErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/insider-one-assessment_internal_utils.HTTPErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "insider-one-assessment_internal_dto_request.CreateTemplateRequest": {
            "type": "object",
            "required": [
                "channel",
                "content",
                "name"
            ],
            "properties": {
                "channel": {
                    "enum": [
                        "sms",
                        "email",
                        "push"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/insider-one-assessment_internal_model.NotificationChannel"
                        }
                    ]
                },
                "content": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "insider-one-assessment_internal_dto_request.UpdateTemplateRequest": {
            "type": "object",
            "properties": {
                "channel": {
                    "enum": [
                        "sms",
                        "email",
                        "push"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/insider-one-assessment_internal_model.NotificationChannel"
                        }
                    ]
                },
                "content": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "insider-one-assessment_internal_dto_resource.TemplateResource": {
            "type": "object",
            "properties": {
                "channel": {
                    "$ref": "#/definitions/insider-one-assessment_internal_model.NotificationChannel"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "insider-one-assessment_internal_dto_resource.TemplateResponse": {
            "type": "object",
            "properties": {
                "template": {
                    "$ref": "#/definitions/insider-one-assessment_internal_dto_resource.TemplateResource"
                }
            }
        },
        "insider-one-assessment_internal_dto_resource.TemplatesResponse": {
            "type": "object",
            "properties": {
                "templates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/insider-one-assessment_internal_dto_resource.TemplateResource"
                    }
                }
            }
        },
        "insider-one-assessment_internal_model.NotificationChannel": {
            "type": "string",
            "enum": [
                "sms",
                "email",
                "push"
            ],
            "x-enum-varnames": [
                "ChannelSMS",
                "ChannelEmail",
                "ChannelPush"
            ]
        },
...
        "insider-one-assessment_internal_utils.ErrorSchema": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "insider-one-assessment_internal_utils.HTTPErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/insider-one-assessment_internal_utils.ErrorSchema"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
...
    }
}